refactor(app): Migrate dataset components and global CSS into app package
