+++ packages/openneuro-app/src/scripts/dataset/components/ModalitiesMetaDataBlock.tsx
@@ -1,6 +1,5 @@
 import React from "react"
 import { Link } from "react-router-dom"
-import Markdown from "markdown-to-jsx"
 
 export interface ModalitiesMetaDataBlockProps {
   items: string[]
