+++ packages/openneuro-app/src/scripts/scss/dataset/dataset-page.scss
@@ -1,4 +1,4 @@
-@import '../scss/variables';
+@import '../variables.scss';
 .dataset {
   .dataset-header {
     h1 {
