+++ packages/openneuro-app/src/scripts/scss/dataset/dataset-tool.scss
@@ -1,4 +1,4 @@
-@import '../scss/variables';
+@import '../variables.scss';
 //share table
 .download-link {
   .return-link {
