+++ packages/openneuro-app/src/scripts/scss/dataset/version-dropdown.scss
@@ -1,4 +1,4 @@
-@import '../scss/variables';
+@import '../variables.scss';
 
 .version-dropdown {
   .toggle {
