+++ packages/openneuro-app/src/scripts/scss/index.scss
@@ -1,0 +1,6 @@
+@import './dataset/comments.scss';
+@import './dataset/dataset-page.scss';
+@import './dataset/dataset-tool.scss';
+@import './dataset/dropdown.scss';
+@import './dataset/validation.scss';
+@import './dataset/version-dropdown.scss';
