--- packages/openneuro-components/src/dataset/index.ts
@@ -1,29 +1,0 @@
-import { MetaDataBlock } from "./MetaDataBlock"
-import { MetaDataListBlock } from "./MetaDataListBlock"
-import { ModalitiesMetaDataBlock } from "./ModalitiesMetaDataBlock"
-import { NemarButton } from "./NemarButton"
-import { BrainLifeButton } from "./BrainLifeButton"
-import { ValidationBlock } from "./ValidationBlock"
-import { CloneDropdown } from "./CloneDropdown"
-import { DatasetHeader } from "./DatasetHeader"
-import { DatasetAlert } from "./DatasetAlert"
-import { DatasetHeaderMeta } from "./DatasetHeaderMeta"
-import { DatasetGitAccess } from "./DatasetGitAccess"
-import { DatasetTools } from "./DatasetTools"
-import { VersionList } from "./VersionList"
-
-export {
-  BrainLifeButton,
-  CloneDropdown,
-  DatasetAlert,
-  DatasetGitAccess,
-  DatasetHeader,
-  DatasetHeaderMeta,
-  DatasetTools,
-  MetaDataBlock,
-  MetaDataListBlock,
-  ModalitiesMetaDataBlock,
-  NemarButton,
-  ValidationBlock,
-  VersionList,
-}
