+++ packages/openneuro-components/src/page/page.scss
@@ -36,11 +36,10 @@
 @import '../search-page/search-result.scss';
 @import '../search-page/search-sort.scss';
 @import '../search-page/filters-block.scss';
-@import '../dataset/dataset-page.scss';
-@import '../dataset/dataset-tool.scss';
-@import '../dataset/comments.scss';
-
-@import '../dataset/validation.scss';
+//@import '../dataset/dataset-page.scss';
+//@import '../dataset/dataset-tool.scss';
+//@import '../dataset/comments.scss';
+//@import '../dataset/validation.scss';
 @import '../json-tree/json-tree.scss';
 
 .page {
